smart_alarm/requirementsAlarm.txt
Adafruit-Blinka==6.14.1
adafruit-circuitpython-dht==3.6.2
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==3.15.3
Adafruit-PureIO==1.1.9
asn1crypto==0.24.0
astroid==2.1.0
asttokens==1.1.13
automationhat==0.2.0
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
cachetools==4.2.2
Cap1xxx==0.1.3
certifi==2018.8.24
cffi==1.14.6
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
crcmod==1.7
cryptography==2.6.1
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.0.2
fourletterphat==0.1.0
google-api-core==2.0.1
google-auth==2.1.0
google-cloud==0.34.0
google-cloud-bigquery==2.28.1
google-cloud-bigquery-storage==2.9.1
google-cloud-core==2.1.0
google-cloud-pubsub==2.8.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
gpiozero==1.6.2
grpc-google-iam-v1==0.12.3
grpcio==1.40.0
html5lib==1.0.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==1.3.1
libcst==0.3.21
logilab-common==1.4.2
lxml==4.3.2
mariadb==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mypy==0.670
mypy-extensions==0.4.1
mysql-connector-python==8.0.26
numpy==1.16.2
oauthlib==2.1.0
olefile==0.46
opencv-python==4.5.3.56
packaging==21.0
pantilthat==0.0.7
parso==0.3.1
pexpect==4.6.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.78
Pillow==5.4.1
pkg-resources==0.0.0
proto-plus==1.19.0
protobuf==3.18.0
psutil==5.5.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pyftdi==0.53.3
pygame==1.9.6
pyparsing==2.4.7
pyserial==3.4
pyusb==1.2.1
PyYAML==5.4.1
requests==2.21.0
rpi-lcd==0.0.2
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
RPLCD==1.3.0
rsa==4.7.2
schedule==1.1.0
SecretStorage==2.3.1
six==1.12.0
smbus==1.1.post2
sn3218==1.2.7
soupsieve==1.8
sysv-ipc==1.1.0
typed-ast==1.3.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
